+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Collecting options for checkbox list and radio list and report if some non-existing option is used
 - Tip for error message "Latte template xxx.latte was not analysed"
 - Tip for standalone templates
+- Feature Transform dynamic form controls to "dynamic" string (control with name $record->id will be transformed to "$record->id") (Turn this feature with parameter `latte.features.transformDynamicFormControlNamesToString: true`)
 
 ### Fixed
 - `If condition is always true` for CheckboxList::getLabelPart(), CheckboxList::getControlPart(), RadioList::getLabelPart() and RadioList::getControlPart()